+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # MiniReelinator
 
 ## v1.9.0 (March 3, 2015)
+* *[v1.9.0.rc9]*
+  * [FIX]: Fix for an issue that caused a new Sponsored Card/MiniReel
+    not to copy the links from a campaign until the campaign was saved
+
 * *[v1.9.0.rc8]*
   * Split Creatives tab into separate Sponsored Cards and Sponsored
     MiniReels tabs